analysis-service/requirements.txt
kafka-python-ng==2.2.3
scikit-learn==1.4.2
numpy==1.26.4
pandas==2.2.1
pytest==8.1.1
